+++ ft_operator.c
@@ -1,26 +1,26 @@
 #include "ft_fdf.h"
 
-void	ft_op_add(float *a, float *b)
+void	ft_op_add(double *a, double *b)
 {
 	*a += *b;
 }
 
-void	ft_op_sub(float *a, float *b)
+void	ft_op_sub(double *a, double *b)
 {
 	*a -= *b;
 }
 
-void	ft_op_mul(float *a, float *b)
+void	ft_op_mul(double *a, double *b)
 {
 	*a *= *b;
 }
 
-void	ft_op_div(float *a, float *b)
+void	ft_op_div(double *a, double *b)
 {
 	*a /= *b;
 }
 
-// void	ft_op_mod(float *a, float *b)
+// void	ft_op_mod(double *a, double *b)
 // {
 // 	*a %= (*b);
 // }
